requirements.txt
openai-agents
python-dotenv
websockets
fastapi
elevenlabs
pyaudio
uvicorn